Change to the format of rules
MRUIZ suggested that the non-optional evaluation of all rules is a bit surprising
to a user, so it's best to stop evaulating on the first match, and let power users
control if they want a rule to apply, but continue evaulating more rules.

Also there would be possible non-deterministic behaviour due to Perl key order
randomization feature in future perls.

That made the config into an ARRAY to force order of evaluation, and RULES
to be hashrefs.